Split mixins and managers out of models.py
